+++ cmd/main.go
@@ -1,6 +1,6 @@
 package main
 
 func main() {
-	// cli.RunTUI(agents.CodingAgent, true, agents.NewCodingContext())
-	// cli.RunTUI(agents.CodingAgent, true, agents.NewCodingContext())
+	// type nilcontext struct{}
+	// agent := agents.Agent[nilcontext]{}
 }
